Added grabbing and putting away nozzle
Item in the player model off hand is now always the game boy horror
Item in the player model main hand is now never the gameboy horror nor interact. It also defaults to the flashlight.
The flashlight is not rendered is the hand of the player model if the poltergust is idled on the back.
Fixed invulnerability blink doesn't show up in idle animations.
Switched the arm the model uses for yelling/waving to the left arm (consistent with the game).
Projectiles created by ghosts now take the owner's attack damage.
Fixed Bowser sometimes spins a whole circle.
Flashlight can now shine when the game boy horror is selected.
Rather than hardcoded ghost revealing, damage swapping and custom model data swapping, mirrors now change the model based on tag.mirror (its contents are merged with item root).
Gooigi in the lab now renders the poltergust in the idle pose.
Changed model variant storage so it stores a whole item, not just custom model data (to ensure mirror can be overwritten).

Death animation is no longer instantly triggered; it now waits until the knockback/scare animation is done.
When health is 0, the max health will now instantly go back to the max.
Improved un-gooigifying poltergust so the beta mansion no longer needs to overwrite the function